Style: 코드 정리
