+++ src/components/bithumb/Bithumb.type.ts
@@ -1,0 +1,18 @@
+export interface ITicker {
+  volumePower: string;
+  chgAmt: string;
+  chgRate: string;
+  prevClosePrice: string;
+  buyVolume: string;
+  sellVolume: string;
+  volume: string;
+  value: string;
+  highPrice: string;
+  lowPrice: string;
+  closePrice: string;
+  openPrice: string;
+  time: string;
+  date: string;
+  tickType: string;
+  symbol: string;
+}
